putting new concordance in new file
