+++ src/rard/research/models/__init__.py
@@ -2,6 +2,7 @@ from .antiquarian import Antiquarian, AntiquarianConcordance
 from .bibliography import BibliographyItem
 from .citing_work import CitingAuthor, CitingWork
 from .comment import Comment
+from .concordance import ConcordanceModel, Edition, PartIdentifier
 from .fragment import AnonymousFragment, AnonymousTopicLink, Fragment, TopicLink
 from .history import HistoricalRecordLog
 from .image import Image
@@ -40,4 +41,7 @@ __all__ = [
     "Work",
     "Book",
     "PublicCommentaryMentions",
+    "ConcordanceModel",
+    "Edition",
+    "PartIdentifier",
 ]
