+++ src/rard/research/models/original_text.py
@@ -146,36 +146,6 @@ class Concordance(HistoryModelMixin, BaseModel):
         return "%s: %s" % (self.source, self.identifier)
 
 
-class AntWorkVolIdentifier(HistoryModelMixin, BaseModel):
-    edition = models.ForignKey("Edition")
-    value = models.CharField(max_length=128, blank=False)
-
-
-class Edition(HistoryModelMixin, BaseModel):
-    edition_identifier = models.CharField(max_length=128, blank=False)
-
-    ant_work_identifier = models.ForeignKey("AntWorkVolIdentifier")
-
-    display_order = models.CharField(max_length=30, blank=True)
-
-
-class ConcordanceModel(HistoryModelMixin, BaseModel):
-    edition = models.ForignKey("Edition")
-
-    TYPE_CHOICES = {
-        "F": "Fragment",
-        "T": "Testimonium",
-        "App": "Appendix",
-        "P.": "Pagination",
-        "NA": "N/A",
-    }
-    content_type = models.CharField(max_length=3, choices=TYPE_CHOICES)
-    reference = models.CharField(max_length=15, blank=True)
-    concordance_order = models.CharField(
-        max_length=15
-    )  # maybe add a validator so only numbers and . allowed?
-
-
 class Translation(HistoryModelMixin, BaseModel):
     history = HistoricalRecords()
